locationStateをLocationObject専用に
