+++ src/components/Permitted.tsx
@@ -20,6 +20,7 @@ import {
   parenthesisRegexp,
 } from '../constants'
 import useAppleWatch from '../hooks/useAppleWatch'
+import { useBadAccuracy } from '../hooks/useBadAccuracy'
 import useCachedInitAnonymousUser from '../hooks/useCachedAnonymousUser'
 import useCheckStoreVersion from '../hooks/useCheckStoreVersion'
 import useConnectivity from '../hooks/useConnectivity'
@@ -77,7 +78,7 @@ const PermittedLayout: React.FC<Props> = ({ children }: Props) => {
   const [longPressNoticeDismissed, setLongPressNoticeDismissed] = useState(true)
 
   const { selectedBound } = useRecoilValue(stationState)
-  const { location, badAccuracy } = useRecoilValue(locationState)
+  const location = useRecoilValue(locationState)
   const setTheme = useSetRecoilState(themeState)
   const [{ autoModeEnabled }, setNavigation] = useRecoilState(navigationState)
   const setSpeech = useSetRecoilState(speechState)
@@ -88,6 +89,7 @@ const PermittedLayout: React.FC<Props> = ({ children }: Props) => {
   const [screenShotBase64, setScreenShotBase64] = useState('')
   const [screenshotTaken, setScreenshotTaken] = useState(false)
   const { subscribing } = useRecoilValue(mirroringShareState)
+  const badAccuracy = useBadAccuracy()
 
   useCheckStoreVersion()
   useAppleWatch()
