+++ src/hooks/useAndroidWearable.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useMemo } from 'react'
 import { Platform } from 'react-native'
 import { useRecoilValue } from 'recoil'
-import locationState from '../store/atoms/location'
 import stationState from '../store/atoms/station'
 import { currentStationSelector } from '../store/selectors/currentStation'
 import getIsPass from '../utils/isPass'
 import sendStationInfoToWatch from '../utils/native/android/wearableModule'
+import { useBadAccuracy } from './useBadAccuracy'
 import useIsNextLastStop from './useIsNextLastStop'
 import { useNextStation } from './useNextStation'
 import { useNumbering } from './useNumbering'
@@ -13,13 +13,13 @@ import { useStoppingState } from './useStoppingState'
 
 const useAndroidWearable = (): void => {
   const { arrived } = useRecoilValue(stationState)
-  const { badAccuracy } = useRecoilValue(locationState)
   const currentStation = useRecoilValue(currentStationSelector({}))
 
   const nextStation = useNextStation()
   const stoppingState = useStoppingState()
   const [currentNumbering] = useNumbering()
   const isNextLastStop = useIsNextLastStop()
+  const badAccuracy = useBadAccuracy()
 
   const station = useMemo(
     () =>
@@ -49,10 +49,10 @@ const useAndroidWearable = (): void => {
       }
     })()
   }, [
-    station,
-    currentNumbering?.stationNumber,
     badAccuracy,
+    currentNumbering?.stationNumber,
     isNextLastStop,
+    station,
     stoppingState,
   ])
 }
