+++ src/screens/Main.tsx
@@ -24,7 +24,6 @@ import TransfersYamanote from '../components/TransfersYamanote'
 import TypeChangeNotify from '../components/TypeChangeNotify'
 import { ASYNC_STORAGE_KEYS, LOCATION_TASK_NAME } from '../constants'
 import useAutoMode from '../hooks/useAutoMode'
-import useDetectBadAccuracy from '../hooks/useDetectBadAccuracy'
 import { useLoopLine } from '../hooks/useLoopLine'
 import useNextOperatorTrainTypeIsDifferent from '../hooks/useNextOperatorTrainTypeIsDifferent'
 import { useNextStation } from '../hooks/useNextStation'
@@ -121,8 +120,8 @@ const MainScreen: React.FC = () => {
   const setBGLocation = useCallback(
     (location: LocationObject) =>
       setLocation((prev) => {
-        if (prev.location?.timestamp !== location.timestamp) {
-          return { ...prev, location }
+        if (prev?.timestamp !== location.timestamp) {
+          return location
         }
         return prev
       }),
@@ -203,7 +202,6 @@ const MainScreen: React.FC = () => {
   useRefreshLeftStations()
   useRefreshStation()
   useKeepAwake()
-  useDetectBadAccuracy()
 
   const handleBackButtonPress = useResetMainState()
   const { pause: pauseBottomTimer } = useUpdateBottomState()
