+++ src/screens/SelectLine.tsx
@@ -52,7 +52,7 @@ const styles = StyleSheet.create({
 })
 
 const SelectLineScreen: React.FC = () => {
-  const [{ location }, setLocationState] = useRecoilState(locationState)
+  const [location, setLocationState] = useRecoilState(locationState)
   const setNavigation = useSetRecoilState(navigationState)
   const setLineState = useSetRecoilState(lineState)
   const fetchStationFunc = useFetchNearbyStation()
@@ -74,10 +74,7 @@ const SelectLineScreen: React.FC = () => {
       if (!pos) {
         return
       }
-      setLocationState((prev) => ({
-        ...prev,
-        location: pos,
-      }))
+      setLocationState(pos)
       await fetchStationFunc(pos)
     }
     init()
@@ -188,10 +185,7 @@ const SelectLineScreen: React.FC = () => {
     if (!pos) {
       return
     }
-    setLocationState((prev) => ({
-      ...prev,
-      location: pos,
-    }))
+    setLocationState(pos)
     setNavigation((prev) => ({
       ...prev,
       stationForHeader: null,
