+++ src/store/atoms/location.ts
@@ -2,17 +2,9 @@ import { LocationObject } from 'expo-location'
 import { atom } from 'recoil'
 import { RECOIL_STATES } from '../../constants'
 
-export interface LocationState {
-  location: LocationObject | null
-  badAccuracy: boolean
-}
-
-const locationState = atom<LocationState>({
+const locationState = atom<LocationObject | null>({
   key: RECOIL_STATES.location,
-  default: {
-    location: null,
-    badAccuracy: false,
-  },
+  default: null,
 })
 
 export default locationState
